Adding user information into the others requests

+++ src/tasks/tasks.service.ts
@@ -15,12 +15,14 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
 
     if (!found) {
       throw new NotFoundException(`Task with id ${id} not found!`);
@@ -32,16 +34,16 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const deleted = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const deleted = await this.taskRepository.delete({ id, userId: user.id });
 
     if (deleted.affected === 0) {
       throw new NotFoundException(`Task with id ${id} not found!`);
     }
   }
 
-  async updateTask(updateTaskDto: UpdateTaskDto): Promise<Task> {
-    const task = await this.getTaskById(updateTaskDto.id);
+  async updateTask(updateTaskDto: UpdateTaskDto, user: User): Promise<Task> {
+    const task = await this.getTaskById(updateTaskDto.id, user);
 
     if (!task) {
       throw new NotFoundException(
@@ -56,8 +58,12 @@ export class TasksService {
     return task;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
 
     if (!task) {
       throw new NotFoundException(`Task with id ${id} not found!`);
